__mytech_archive/docker/AISuite/requirements.AISuite.txt
# Web Frameworks
setproctitle
requests
watchdog  # Library for monitoring file system events --- #!speeds up streamlit
streamlit
# Async Libraries
asyncio
aiofiles

# Common Python Libraries
python-dotenv
faker
nltk
spacy
pandas
numpy
setproctitle


# Database Libraries
psycopg2-binary
asyncpg
sqlalchemy-utils  # Various utility functions for SQLAlchemy
chromadb


# Visual Studio Code
ptvsd
debugpy

langchain # Python client for LangChain API
langchainhub # Python client for LangChainHub API
langchain-community # Python client for LangChain Community API
langsmith # Python client for LangChain AI API
langchain_openai
ollama
openai

playwright
html2text
tldextract


streamlit-elements
streamlit_extras
streamlit_code_editor
streamlit-aggrid
